story/NP-01 Some refactoring
