+++ src/course/models/course.entity.ts
@@ -1,7 +1,8 @@
 import { ValidationTypes } from "class-validator";
+import { VideoEntity } from "../../videos/models/video.entity";
 import { Entity, Column, PrimaryGeneratedColumn, OneToMany, JoinColumn } from "typeorm";
 import { ArticleEntity } from "./article.entity";
-import { VideoEntity } from "./video.entity";
+
 
 @Entity('course')
 export class CourseEntity {
