+++ src/course/models/video.entity.ts
@@ -1,22 +1,22 @@
-import { Entity, Column, PrimaryGeneratedColumn, OneToMany, ManyToMany, ManyToOne, JoinColumn } from "typeorm";
-import { CourseEntity } from "./course.entity";
+// import { Entity, Column, PrimaryGeneratedColumn, OneToMany, ManyToMany, ManyToOne, JoinColumn } from "typeorm";
+// import { CourseEntity } from "../course/models/course.entity";
 
 
-@Entity({ name: 'videos' })
-export class VideoEntity {
-    @PrimaryGeneratedColumn()
-    id: string;
+// @Entity({ name: 'videos' })
+// export class VideoEntity {
+//     @PrimaryGeneratedColumn()
+//     id: string;
 
-    @Column()
-    name: string;
+//     @Column()
+//     name: string;
 
-    @Column()
-    title: string;
+//     @Column()
+//     title: string;
 
-    @Column()
-    url: 'text';
+//     @Column()
+//     url: 'text';
 
-    @ManyToOne(() => CourseEntity, (course: CourseEntity)=> course.videos)
-    @JoinColumn({ name: 'course_id' })
-    course: CourseEntity;// this defines the property name
-}
+//     @ManyToOne(() => CourseEntity, (course: CourseEntity)=> course.videos)
+//     @JoinColumn({ name: 'course_id' })
+//     course: CourseEntity;// this defines the property name
+// }
