+++ src/videos/models/video.entity.ts
@@ -1,0 +1,24 @@
+import { CourseEntity } from "../../course/models/course.entity";
+import { Entity, Column, PrimaryGeneratedColumn, OneToMany, ManyToMany, ManyToOne, JoinColumn } from "typeorm";
+
+
+
+@Entity({ name: 'video' })
+export class VideoEntity {
+    @PrimaryGeneratedColumn()
+    id: string;
+
+    @Column()
+    name: string;
+
+    @Column()
+    title: string;
+
+    @Column()
+    url: 'text';
+
+    @ManyToOne(() => CourseEntity, (course: CourseEntity)=> course.videos)
+    @JoinColumn({ name: 'course_id' })
+    course: CourseEntity;// this defines the property name
+}
+
